IDEX-2972: move extension server controller to machine-gwt-client
