+++ plugin-machine/che-plugin-machine-ext-client/src/main/java/org/eclipse/che/ide/extension/machine/client/MachineExtension.java
@@ -28,6 +28,7 @@ import org.eclipse.che.ide.extension.machine.client.actions.SelectCommandComboBo
 import org.eclipse.che.ide.extension.machine.client.actions.SwitchPerspectiveAction;
 import org.eclipse.che.ide.extension.machine.client.machine.console.ClearConsoleAction;
 import org.eclipse.che.ide.extension.machine.client.machine.console.MachineConsoleToolbar;
+import org.eclipse.che.ide.extension.machine.client.machine.extserver.ExtServerStateNotifier;
 import org.eclipse.che.ide.extension.machine.client.machine.extserver.ProjectApiComponentInitializer;
 import org.eclipse.che.ide.extension.machine.client.outputspanel.OutputsContainerPresenter;
 import org.eclipse.che.ide.ui.toolbar.ToolbarPresenter;
@@ -64,6 +65,8 @@ public class MachineExtension {
                                 SelectCommandComboBoxAction selectCommandAction,
                                 EditCommandsAction editCommandsAction,
                                 CreateMachineAction createMachine,
+                                //Don't remove it's need for initialization
+                                ExtServerStateNotifier extServerStateNotifier,
                                 RestartMachineAction restartMachine,
                                 DestroyMachineAction destroyMachine,
                                 SwitchPerspectiveAction switchPerspectiveAction) {
