--- plugin-machine/che-plugin-machine-ext-client/src/main/java/org/eclipse/che/ide/extension/machine/client/machine/events/ExtServerStateHandler.java
@@ -1,37 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2012-2015 Codenvy, S.A.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *   Codenvy, S.A. - initial API and implementation
- *******************************************************************************/
-package org.eclipse.che.ide.extension.machine.client.machine.events;
-
-import com.google.gwt.event.shared.EventHandler;
-
-/**
- * A handler for handling {@link ExtServerStateEvent}.
- *
- * @author Roman Nikitenko
- */
-public interface ExtServerStateHandler extends EventHandler {
-
-    /**
-     * Called when extension server has been started.
-     *
-     * @param event
-     *         the fired {@link ExtServerStateEvent}
-     */
-    void onExtServerStarted(ExtServerStateEvent event);
-
-    /**
-     * Called when extension server has been stopped.
-     *
-     * @param event
-     *         the fired {@link ExtServerStateEvent}
-     */
-    void onExtServerStopped(ExtServerStateEvent event);
-}
