--- plugin-machine/che-plugin-machine-ext-client/src/main/java/org/eclipse/che/ide/extension/machine/client/machine/extserver/ExtServerState.java
@@ -1,20 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2012-2015 Codenvy, S.A.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *   Codenvy, S.A. - initial API and implementation
- *******************************************************************************/
-package org.eclipse.che.ide.extension.machine.client.machine.extserver;
-
-/**
- * Describes state of a extension server
- *
- * @author Roman Nikitenko.
- */
-public enum ExtServerState {
-    STARTED, STOPPED
-}
