+++ plugin-machine/che-plugin-machine-ext-client/src/main/java/org/eclipse/che/ide/extension/machine/client/machine/extserver/ProjectApiComponentInitializer.java
@@ -19,8 +19,8 @@ import com.google.web.bindery.event.shared.EventBus;
 import org.eclipse.che.ide.bootstrap.ProjectTemplatesComponent;
 import org.eclipse.che.ide.bootstrap.ProjectTypeComponent;
 import org.eclipse.che.ide.core.Component;
-import org.eclipse.che.ide.extension.machine.client.machine.events.ExtServerStateEvent;
-import org.eclipse.che.ide.extension.machine.client.machine.events.ExtServerStateHandler;
+import org.eclipse.che.api.machine.gwt.client.ExtServerStateEvent;
+import org.eclipse.che.api.machine.gwt.client.ExtServerStateHandler;
 import org.eclipse.che.ide.util.loging.Log;
 
 /**
